Add Category & Add Expenses

+++ src/main/model/Income.java
@@ -3,16 +3,16 @@ package main.model;
 import java.util.Date;
 
 public class Income {
-    private int value;
+    private float value;
     private Date date;
     public static int totalBalance=0;
 
-    public Income(int value, Date date) {
+    public Income(float value, Date date) {
         this.value = value;
         this.date = date;
     }
 
-    public int getValue() {
+    public float getValue() {
         return value;
     }
